#16 fix actuator integration bug
