+++ client/lookout-sofa-boot-starter/pom.xml
@@ -25,15 +25,15 @@
         </developer>
     </developers>
     <dependencyManagement>
-      <dependencies>
-         <dependency>
-            <groupId>org.springframework.boot</groupId>
-            <artifactId>spring-boot-dependencies</artifactId>
-            <version>1.4.2.RELEASE</version>
-            <type>pom</type>
-            <scope>import</scope>
-          </dependency>
-       </dependencies>
+        <dependencies>
+            <dependency>
+                <groupId>org.springframework.boot</groupId>
+                <artifactId>spring-boot-dependencies</artifactId>
+                <version>1.4.2.RELEASE</version>
+                <type>pom</type>
+                <scope>import</scope>
+            </dependency>
+        </dependencies>
     </dependencyManagement>
 
     <dependencies>
