+++ client/lookout-sofa-boot-starter/src/main/java/com/alipay/lookout/starter/autoConfiguration/LookoutAutoConfiguration.java
@@ -26,9 +26,6 @@ import com.alipay.lookout.remote.report.AddressService;
 import com.alipay.lookout.remote.step.LookoutRegistry;
 import com.alipay.lookout.report.MetricObserver;
 import com.alipay.lookout.starter.LookoutClientProperties;
-import com.alipay.lookout.starter.support.actuator.LookoutSpringBootMetricsImpl;
-import com.alipay.lookout.starter.support.actuator.SpringBootActuatorRegistry;
-import com.alipay.lookout.starter.support.reader.LookoutRegistryMetricReader;
 import com.alipay.lookout.starter.support.reg.*;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -37,14 +34,11 @@ import org.springframework.beans.factory.BeanFactory;
 import org.springframework.beans.factory.BeanFactoryAware;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.actuate.autoconfigure.MetricsDropwizardAutoConfiguration;
-import org.springframework.boot.actuate.endpoint.MetricReaderPublicMetrics;
-import org.springframework.boot.actuate.metrics.CounterService;
-import org.springframework.boot.actuate.metrics.GaugeService;
 import org.springframework.boot.autoconfigure.AutoConfigureBefore;
 import org.springframework.boot.autoconfigure.AutoConfigureOrder;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnBean;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -61,7 +55,7 @@ import static com.alipay.lookout.core.config.LookoutConfig.*;
 @AutoConfigureOrder(-100)
 @Configuration
 @EnableConfigurationProperties(LookoutClientProperties.class)
-@AutoConfigureBefore(MetricsDropwizardAutoConfiguration.class)
+@AutoConfigureBefore({ MetricsDropwizardAutoConfiguration.class })
 public class LookoutAutoConfiguration implements BeanFactoryAware {
     private static final Logger                      logger = LoggerFactory
                                                                 .getLogger(LookoutAutoConfiguration.class);
@@ -99,12 +93,6 @@ public class LookoutAutoConfiguration implements BeanFactoryAware {
     }
 
     @Bean
-    @ConditionalOnClass(name = "org.springframework.boot.actuate.metrics.Metric")
-    public SpringBootActuatorRegistryFactory springBootActuatorServerRegistryFactory() {
-        return new SpringBootActuatorRegistryFactory();
-    }
-
-    @Bean
     @ConditionalOnClass(name = "com.alipay.lookout.reg.prometheus.PrometheusRegistry")
     public MetricsRegistryFactory prometheusMetricsRegistryFactory() {
         return new PrometheusMetricsRegistryFactory();
@@ -117,10 +105,12 @@ public class LookoutAutoConfiguration implements BeanFactoryAware {
      *
      * @return
      */
+    @Deprecated
     @Bean
+    @ConditionalOnProperty(prefix = "com.alipay.sofa.lookout", name = "actuator-dropWizard-enabled", havingValue = "true", matchIfMissing = true)
     @ConditionalOnClass(name = { "com.alipay.lookout.dropwizard.metrics.DropWizardMetricsRegistry",
             "com.codahale.metrics.MetricRegistry" })
-    public MetricsRegistryFactory dropWizardMetricsRegistryFactory() {
+    public DropWizardMetricsRegistryFactory dropWizardMetricsRegistryFactory() {
         try {
             com.codahale.metrics.MetricRegistry metricRegistry = beanFactory
                 .getBean(com.codahale.metrics.MetricRegistry.class);
@@ -134,6 +124,13 @@ public class LookoutAutoConfiguration implements BeanFactoryAware {
             logger.debug("no dropwizard service found.");
         }
         return null;
+    }
+
+    @Bean
+    @ConditionalOnMissingBean({ DropWizardMetricsRegistryFactory.class })
+    @ConditionalOnClass(name = "org.springframework.boot.actuate.metrics.Metric")
+    public SpringBootActuatorRegistryFactory springBootActuatorServerRegistryFactory() {
+        return new SpringBootActuatorRegistryFactory();
     }
 
     @Bean
@@ -151,30 +148,6 @@ public class LookoutAutoConfiguration implements BeanFactoryAware {
         lookoutClient.registerExtendedMetrics();
 
         return lookoutClient.getRegistry();
-    }
-
-    @Bean
-    @ConditionalOnMissingBean({ LookoutSpringBootMetricsImpl.class, CounterService.class,
-            GaugeService.class })
-    public LookoutSpringBootMetricsImpl lookoutMetricServices(Registry lookoutMetricRegistry) {
-        logger.info("Spring Boot Metrics binding to SOFALookout Implementation!");
-        return new LookoutSpringBootMetricsImpl(lookoutMetricRegistry);
-    }
-
-    @Bean
-    @ConditionalOnMissingBean
-    @ConditionalOnBean(SpringBootActuatorRegistryFactory.class)
-    public LookoutRegistryMetricReader lookoutRegistryMetricReader(SpringBootActuatorRegistryFactory springBootActuatorRegistryFactory,
-                                                                   LookoutConfig lookoutConfig) {
-        SpringBootActuatorRegistry springBootActuatorRegistry = springBootActuatorRegistryFactory
-            .get(lookoutConfig);
-        return new LookoutRegistryMetricReader(springBootActuatorRegistry);
-    }
-
-    @Bean
-    @ConditionalOnBean(LookoutRegistryMetricReader.class)
-    public MetricReaderPublicMetrics lookoutPublicMetrics(LookoutRegistryMetricReader lookoutRegistryMetricReader) {
-        return new MetricReaderPublicMetrics(lookoutRegistryMetricReader);
     }
 
     protected LookoutConfig buildLookoutConfig(LookoutClientProperties lookoutClientProperties) {
