+++ client/lookout-sofa-boot-starter/src/main/java/com/alipay/lookout/starter/support/actuator/LookoutSpringBootMetricsImpl.java
@@ -35,15 +35,9 @@ import java.util.concurrent.ConcurrentMap;
  */
 public class LookoutSpringBootMetricsImpl implements CounterService, GaugeService {
 
-    public static final String                              LOOKOUT_PREFIX         = "lookout.";
-
-    private static final String                             LOOKOUT_COUNTER_PREFIX = LOOKOUT_PREFIX;
-
-    private static final String                             LOOKOUT_GAUGE_PREFIX   = LOOKOUT_PREFIX;
-
     private final Registry                                  registry;
 
-    private final ConcurrentMap<String, SimpleLookoutGauge> gauges                 = new ConcurrentHashMap<String, SimpleLookoutGauge>();
+    private final ConcurrentMap<String, SimpleLookoutGauge> gauges = new ConcurrentHashMap<String, SimpleLookoutGauge>();
 
     public LookoutSpringBootMetricsImpl(Registry registry) {
         this.registry = registry;
@@ -54,7 +48,6 @@ public class LookoutSpringBootMetricsImpl implements CounterService, GaugeServic
         if (StringUtils.isBlank(metricName)) {
             return;
         }
-        metricName = wrapName(LOOKOUT_COUNTER_PREFIX, metricName);
         Id id = this.registry.createId(metricName);
         Counter counter = this.registry.counter(id);
         counter.inc();
@@ -65,7 +58,6 @@ public class LookoutSpringBootMetricsImpl implements CounterService, GaugeServic
         if (StringUtils.isBlank(metricName)) {
             return;
         }
-        metricName = wrapName(LOOKOUT_COUNTER_PREFIX, metricName);
         Id id = this.registry.createId(metricName);
         Counter counter = this.registry.counter(id);
         counter.dec();
@@ -76,7 +68,6 @@ public class LookoutSpringBootMetricsImpl implements CounterService, GaugeServic
         if (StringUtils.isBlank(metricName)) {
             return;
         }
-        metricName = wrapName(LOOKOUT_COUNTER_PREFIX, metricName);
         Id id = this.registry.createId(metricName);
         this.registry.removeMetric(id);
     }
@@ -86,9 +77,6 @@ public class LookoutSpringBootMetricsImpl implements CounterService, GaugeServic
         if (StringUtils.isBlank(metricName)) {
             return;
         }
-        //name better
-        metricName = wrapName(LOOKOUT_GAUGE_PREFIX, metricName);
-
         SimpleLookoutGauge gauge = this.gauges.get(metricName);
         if (gauge == null) {
             SimpleLookoutGauge newGauge = new SimpleLookoutGauge(value);
@@ -100,23 +88,6 @@ public class LookoutSpringBootMetricsImpl implements CounterService, GaugeServic
             }
         }
         gauge.setValue(value);
-    }
-
-    /***
-     * Construce lookout prefix
-     *
-     * @param metricName metric name
-     * @param prefix lookout prefix
-     * @return unique metric
-     */
-    private String wrapName(String prefix, String metricName) {
-        if (StringUtils.isBlank(metricName)) {
-            throw new RuntimeException("Metric name can't be blank!");
-        }
-        if (metricName.startsWith(prefix)) {
-            return metricName;
-        }
-        return prefix + metricName;
     }
 
     private final static class SimpleLookoutGauge implements Gauge<Double> {
