+++ client/lookout-sofa-boot-starter/src/main/resources/META-INF/spring.factories
@@ -1,6 +1,7 @@
 # Auto Configure
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-com.alipay.lookout.starter.autoConfiguration.LookoutAutoConfiguration
+com.alipay.lookout.starter.autoConfiguration.LookoutAutoConfiguration,\
+com.alipay.lookout.starter.configuration.MetricsActuatorAutoConfiguration
 
 
 # Initializers
