+++ client/lookout-sofa-boot-starter/src/test/resources/application-metrics.properties
@@ -1,1 +1,3 @@
 com.alipay.sofa.lookout.prometheusExporterServerPort=9696
+com.alipay.sofa.lookout.actuator-dropWizard-enabled=false
+
