+++ client/lookout-sofa-boot-starter/src/test/resources/application-reader.properties
@@ -1,1 +1,2 @@
 com.alipay.sofa.lookout.prometheusExporterServerPort=9595
+com.alipay.sofa.lookout.actuator-dropWizard-enabled=false
