+++ client/samples/lookout-client-samples-dropwizard-boot/src/main/resources/application.properties
@@ -1,0 +1,3 @@
+management.security.enabled=false
+spring.application.name=lookout-client-samples-boot
+#com.alipay.sofa.lookout.agent-host-address=localhost
